Update Pork2Sausage.h
In the introductory comments, I have corrected a couple of grammatical errors and changed the reference to the plugin name to the external user-facing name "Pork to Sausage", as the internal source project name is evident from the file name.

My substantive change is to amend code indentation in order to improve clarity and comprehension.  I have also added inline comments on to the end of two #endif statements.